Fix definition of |derived@RealForm|, which was doing linear algebra wrong
It was using the |right_inverse| of a projection map, which when the projection
is applied at the right as injection of the derived coweight lattice into the
original coweight lattice amounts to some projection splitting this injection.
The right thing to do here however is to apply the orthogonal projection onto
the derived rational coweight space, which is done using fundamental coweights.