Clean up translate.at, define wall intersection projector in basic.at
The wall intersection prjector is orthogonal; its definition was motivated by
the false belief that this was what was being done in translate.at, and that
each of these non-commuting projections needed separate computation. However the
shifts in |to_wall| from translate.at are parallel to other walls, and therefore
commute, so the basic idea could be maintained and the projectors from basic.at
not used. But the orthogonal projections may be useful elsewhere, so we leave
the defintion in basic.at in place.

The functions in translate.at were also reorganised. The mentioned shifting to
wall intersections was made more robust by not insisting that each separate
shift be an integer vector, and testing just their sum for this property.

The function |translate_by@(Param,vec)| returns a |Param| (non finalised), while
|T@(Param,ratvec)| returns a (possibly zero) |ParamPol|, and all versions of
|Psi| do so as well. Clients that used to rely on |Psi| returning a |Param| for
a |Param| argument (as it used to) have been modified to either use
|translate_by| instead, or apply |first.param| in cases where this is not
feasible, but the |ParamPol| is known nonetheless to be nonzero.