Repair and clean up translate.at, define wall intersection projector in basic.at
The code for translating to the intersection of walls was assuming it can be
done one wall at the time, which is false. The required orthogonal projection
matrices are now defined in basic.at as |root_span_projector| and it complement
|wall_projector|. For the rest functions in translate.at were reorganised.

The final function |make_maximally_singular| appears to suffer from the same
failing commutation assumption, but was left in with a warning comment.