Further progress on coherent continuation, now incuding theta_induce_irr_coherent and theta_induce_irr_wgraph
The latter is defined at the end of coherent_irreducible.at,
to resolve circularity issues.

If p is a parameter for L in G,

theta_induce_irr(p,G)
theta_induce_irr_coherent(p,G)
theta_induce_irr_wgraph(p,G)

should all give identical results.

theta_induce_irr_wgraph(block,graph,p,G)

is the command which has a chance of being faster; it requires
precocmputing the block *for G* of the induced representation, and its
WGraph. theta_induce_irr_wgraph(p,G) does this on the fly,
so is slow, but this is useful for testing.