Adapt |Free_Abelian_light| to the fact that sometimes zero terms exist in |main|
Since it can happen that adding a term actually cancels an existing term, the
existence of zero terms in then |main| vector has always been a possibility,
(though not one we were aware of). It does not real harm as at worst it
ultimately leads to adding a zero multiple of a |StandardRepr| to an |SR_poly|,
which request will then be gracefully ignored. We can however make the internal
operations of |Free_Abelian_light| hide their existence (until such time as they
can be squeezed out in some combination or flattening operation) by defining
iterators in such a way as to to never point to such zero terms; this involves
both |operator++| and |begin|. The |flatten| method is changed to suppress these
terms, and iterator begin and increment to skip them. By contrast, similar
appearance of zero terms in |recent| is cheap to handle immediately, so there
we erase them right away. Finally, now unsused, we remove the method |snapshot|.