Move rather than copy formulas to tables in (twisted) |Rep_table::deformation|
There use to be two copies made at the end of computation: one upon storage, and
another upon returning that stored value. The first is now eliminated; the
second copy cannot be avoided (by returning a reference) as long as some cases
return a non-stored value, as they do. However the second copy is not useless
because it can later be pilfered from when adding the result to other formulas.