Implement forgotten polynomial sharing shortcut in |kl::KL_table::swallow|
One motivation for sharing a polynomial table across blocks for a same real form
is that it makes polynomial index renumbering unnecessary when incorporating an
old partial block into a new larger one. The test for this case, which was made
the standard (due to passing the hash table stored in |Rep_table| to (nearly)
all blocks constructed within the context of this real form), was however
missing in |kl:KL_table::swallow|; interestingly, it was already made in the
corresponding |ext_kl:KL_table::swallow|. The latter was nonetheless slightly
modified by making the test for hash tables sharing the same underlying vector
of polynomials, which is the relevant property, although in practice this
happens if and only if the hash tables themselves are the same object.