Make ratvec division/multiplication be scalar a free function rather than method
The is to allow the rational vector to be passed by value, so that a copy is
only made in case an lvalue is supplied, in which case it is necessary anyway.
A downside is that the scalar must now match the template type exactly (with an
independent (integral) template type this would match too greedily), forcing the
use of length indication/conversion in the operand. Maybe not an ideal situation.