Provide additional offset argument to |StandardRepr| reconstruction, zero for now
When in the future |StandardReprMod| will implements a coarser equality and
corresponding hashing, we must cater for a difference between the values
$\gamma-\lambda$ or $\gamma-\lambda+\rho$ of the supplied and found element when
looking up a block. This difference must then be transmitted to the point where
|StandardReprMod| must be reconstructed from block elements. The new method
|Rep_table::offset| computes the difference, and new arguments are added to
existing methods to transport the value found to where it is needed. So the
infrastructure for the transmission is in place, but currently the value passed
will be zero since matching of |StandardReprMod| still means exact equality.

As implemented a uniform shift to $\gamma-\lambda$, independent of the
involution, will only work if it is entirely neutral to the block construction,
which essentially means it must be orthogonal to the integral system. The
offset method should ensure this property, which might involve changing
representative for the equivalence class of $\lambda$ at the current involution;
this is currently the case because it always returns zero, but with the future
change this will require special attention.