Fix |common_context| slicing problem by no longer letting it be derived
The temptation to use a reference to the base |Rep_context| of |common_context|
is great, but |common_context| values are shorter lived thatn for instance the
|common_block| values they hallped to construct, so this could (and did) lead to
dangling references. To eradicate the problem, make |common_context|, rather
than creating its own on construction, hold a reference to a longer lived
|Rep_context|, which then may validly be copied into a |common_block|. This
requires adding some methods to simulate derivation, and other logistic changes.