Make |Reduced_param| ignore variations in $(1-\theta)X^*$
Basically we apply |codec.in| to the evaluation to bring it into coordinates
that line up with the (reduced) image of $(1-\theta)X^*$, and then reduce
coordinates modulo the entris of |codec.diagonal|. For this it is essential that
those entries be positive, which was not ensured for the first one. Flipping the
sign also requires flipping the first row of |codec.out| to compensate. All in
all this makes the trivial of SO(6,3) compute correctly, but SO(5,4) still bugs