Make blocks have the possibility of multiple extended blocks
The grouping of different (modular) parameters under a same |Reduced_param|
ensures isomorphic blocks, but the extended block construction may differ
between so identified blocks: the set of twist-fixed points differs for shifts
that are not $\delta$-fixed, and more subtly the signs for default-constructed
extended parameters can be affected even by shift that are $\delta$-fixed.

Therefore we allow more than one extended block to be associated to a given
block, the different ones being identified by the (normalised) value of
|lambda_gamma| at the first block element (whether it detemines an extended
block elements or not). The implementation generates the extended blocks by
changing the |gamma_lambda| values in the block itself, in a way that should not
affect other (non extended) uses of the block (such changes may happen enyway
when a partial block gets swallowed into a larger one with different
|gamma_lambda| values, so this ought to be OK).