Add method |common_block::shift| that allows changing representatives
As long as this is done at a quiet moment, globally shifting the block element
representatives in their |gamma_lambda| component should not affect its
validity; it just changes the block into one that could have been constructed
directly from a different initial (modular) parameter. The utility is that this
is a relatively easy way to ensure that the block appears exactly as if it was
specifically constructed for a particular parameter. This is notably useful for
extended block construction: we have a twist-stable actual parameter, and we can
(temporarily) transform the common block so that its element matching that
parameter becomes twist-stable, which the extended block construction assumes.

Applying the shift requires the |common_block::srm_hash| to be reconstructed.