Change interpretation of |RatWeight| accompanying extended blocks: now a shift
This used to be a value of |gamma_lambda| at element 0, but those values are not
useful for the whole block: the have been modularly reduced and situations arise
where a shift does not affect element 0 but it does affect other elements.
Therefore we now instead record a shift that the unexended block needs to
undergo in order to be aligned with this extended block for interpreting the
latter. An implication of this change is that any shift actually applied to the
block (using |common_block::shift|) should also make a correction to these
rational vectors associated to each extended block stored in its |extended|.