Activate extended block storage and KL table reuse (not good for large cases)
Extended blocks are now generated and stored in the context of a |common_block|,
and since shifting within a class of modular parameters combined by using
|Reduced_param| may alter the structure of the extended block, many of them are
combined in the |common_block::extended| list. Currently we always apply a shift
to align the block before creating an extended block, and this can imply a lot
of shifting to a new block when integrating extended blocks from smaller (or
overlapping) existing blocks. Currently this causes rational arithmetic overflow
in large cases like E6_s, so an alternative mechanism to shifting should
probably be implemented and hopefully will avoid such problems.