Remove storage for involution-indexed |codec| data in each |integral_data_item|
This turns out to be a lot of memory (in large examples) spent for a relatively
small gain in computing time, since there are many integral-involution
combinations for a given inner class (which is where these are stored). For now
we do keep the |integral_data_item| data themselves, which only depend on the
integral system, and the |InvolutionTable| data, which depend only on the
involution and have been stored for a very long time, since individually these
provide storage uses that are relatively cost-effective.