Fix stupid implementation of |RationalVector::integer_diff|; clean up arithmetic
When two |RationalVector| values are supposed to have integer difference, it is
stupid to multiply their denominators which are probably equal and certainly
should become so after normalization; avoiding this also avoids overflow risk.

In implemneting the change it became apparent that considering |normalize| to be
a manipulator is somewhat costly: it requires making a copy in some cases, while
the change made can only be beneficial. Therefore we redefined a new version
that is declared and accessor. In passing we also inspected the implementation
of this method, which could make beter use of the |arithmetic::gcd| routine that
is asymmetric in its argument types (we now avoid additional tests or casts).