Revise rational vector arithmetic; do more in place, normalize after additive ops
For additive operations on rational vectors, copies can mostly be avoided,
except for producing a result when neither operand can be used as working
storage. The old code produced as denominator the least common multiple of those
of the operands, which if repeated can lead to uncontrollled growth; therefore
the additive operations now normalize before return, to squeeze out common
factors that might have been produced in the linear combination of numerators.