Review more rational vector arithmetic in repr.cpp, avoiding premature narrowing
The mentioned narrowing took place in the method |Rep_context::sr_gamma|, where
it only affected the height computation. We took the occasions to also change
the implementation of |Rep_context::lambda_rho| to follow the same pattern.

In passing, some more variants of addition/subtraction of rational with integer
vector were added, to avoid a possible needless copy of an rvalue left operand.