Further bug fixes to associated_variety_ann_* and av_maximal_ideal
The main bug was that the gamma given in av_maximal_ideal
in the coordinates of some given RootDatum rd, but this rd wasn't
passed. We are given a SpringerTable st, and st.root_datum is of the
same type as rd, but with different coordinates.

Solution: pass the original rd_in, then immediately convert gamma_in
to gamma in coordinates for st.root_datum. Then rd_in is no longer
used.