Make the |deform| builtin function finalize its argument, avoids possible crash
The |Repr_table::deformation_terms| method implementing this function requires
(and |assert|s) the argument to be final, so the wrapper should either test or
ensure that this condition is met, neither of which were the case. It is now
ensure by calling |finals_for| and summing over the contributions so obtained.