Redefine hash table requirements to allow avoiding any |Entry| copy construction
Inside the |reconstruct| method (only), the |hashCode| needs to be computed from
a |Pooltype::const_reference|, and in order to call that method from |Entry|,
one was construction from the reference on the spot. That means a temporary copy
is explicitly constructed, which is quite useless in the common case that the
|const_reference| is just |const Entry&|. This is now resolved by instead
calling a free function |hash_code| for the reference (and the modulus), which a
template definition will convert into a |hashCode| method call, without any copy
being made. When |const_reference| differs from |const Entry&|, as is currently
the case for |PolEntry| and |SafePolEntry| in polynomials.h, an explicit
definition of |hash_code| can be provided that avoid calling an |Entry| method.
In most cases (and for polynomials in particular) no copy is needed here either.

Finally this is also a preparation for allowing hashable types (notably
|K_type_poly|) to have move semantics only, not having a copy constructor at all.