Reorganise printing of |StandardRepr| and friends to be based in the library
This involes also |Split_integer|, |SR_Poly|, |K_type| and |K_type_poly|, and is
done using free functions (not |operator<<|) defined in basic_io; the similar
methods |Rep_context::print| were not used, and have been removed.

This should make it easier to study the functioning of the deformation algorithm
either adding diagnostic output calls, or in the debugger.