Fix two silly errors in alcove.at; however |promote_in_alcove| seems doomed
The |promote_in_alcove| was returning its initial argument rather than the
"corrected" value, provided its tests were passed, and |in_same_alcove| was
missing a division by 2 leading to incorrect call of |is_parity|. These error
were fixed. However, the case in SO(4,4) that brought the latter error to light
also showed that the idea of finding a canonical (and simple) point in each
alcove by moving to the barycentre of the intersection of "allowed" walls will be
hard to obtain: for SO(4,4) parameter(KGB(x,24),[2,1,1,0],[5/3,4/3,1/3,0]), the
freedom within the walls one is already on is a triangle, all of which sides
belong to the alcove, but none of whose vertices do.