Implement |improve@(Param,int)| from alcoveRep.at in new C++ library module
Since repr.cpp was getting huge, a new pair of files alcoves.{h,cpp} was
created, for now with mainly the large free function |make_multiple_integral|.
Functionality is not yet complete for insertion into the deformation algorithm.

In the fututre some alcove related stuff might be moved from repr.cpp to this
module as well, for easier navigation.