Correct and improve |wall_set|: don't insist that the root datum prefer coroots
Though it is pleasant to think that coroots are ordered by increasing additive
level, as is the case when it prefers coroots, the validity of the wall
selecting method should not depend on this. Indeed, the case where this was
thought to be important is when two coroots have the same evaluation fractional
part because they differ by an integral coroot, and the additively lower one
should be taken if the positive integral coroot is already in the wall set;
however since we now filter away any coroots that can be lowered by subtracting
already selected wall coroots, in this situation only the lower one should
remain a candidate anyway.

Apart from that, changing the root datum to one that prefers coroots would give
the obligation to renumber coroots afterwards, which was not done, so the old
code actually was giving erroneous results in those cases; this is now correct.

Finally we made an unrelated fix to the |strong_components| function, affecting
only the unlikely case that it was being called from a place where the result
would be unused; we had forgotten to not produce any result in such cases.