Make |splice| return proper iterator in all no-op (iterator coincidence) cases
When some iterators in a |splice| call coincide, the requested cut-and-paste
operation may result in the initial situation. This was handled by an early
cop-out, but when that happens, |splice| should return the iterator to the range
end, and this was not always the case. In practice this change won't help much,
since in these cases the client might still be confused by the fact that the
iterator that pointed to the range before it was "spliced away" is still
pointing to it afterwards; therefore client must still be aware of, and probably
circumvent, such cases where |splice| is a no-op.