Change internal |alcove_center| to preserve integrality of coroots
Previously |alcove_center| would move to walls if they are for positive coroots;
the asymmetry this caused for complex coroots would lead to changes in the
$\theta$-fixed part of an infinitiesmal character, which is not allowed. The
current commit will not move to any wall one is not already on, and create
symmetry with respect to the other walls, so it should hopefull fix the issue.

The code will throw a runtime error when this assumption turns out to be false.