Small changes in include directives to always include Atlas.h sufficiently early
Thus all non-utility Atlas files will depend during compilation only on the
definitions of Atlas.h, not on the forward files in the utilities directory
whose contents they duplicate. This is just for fixing the rules followed; it
does not make it possible to link when inconsistencies exist between Atlas.h and
those forward files, since the object files in the utility directory itself will
be compiled without any use of Atlas.h, so object files would mismatch.