Shuffle template code from matrix.cpp to matrix.h, making instantiation implicit
For all templated methods, and most templated free functions, we provide
definitions in matrix.h so that instantiation will be implicit. By removing the
need to explicitly instatiate in matrix.cpp, we avoid the problem that the
number of required instantiations is dependent on symbolic type definitions,
with the risk of either repeated or missing instantiations.