Change implementation of |simple_list| and |sl_list|, avoiding double allocators
As a container class these types stored (were derived from) their allocator
type, but the contained |std::unique_ptr| is also derived from its deleter,
which for proper functioning also stores (is derived from) the allocator type.
This combines to give double storage of allocator objects; in current practice
they are |std::allocator| instances which are stateless (no stored data), but
the fact that there were two of them prevented the Empty Base Optimisation.

We now refrain from storing the allocator directly, obtaining it instead from
the stored |std::unique_ptr| deleter, making sure (which was not the case
before) that the allocator stored there is in fact the one that was passed to
list constructor (in case it is statefull). This requires a fairly thorough
revision of sl_list.h (many constructors need changes) and the occasion has been
taken to clean up unrelated suspicious code, notably in the unused part of the
|swap| methods dealing with incompatible and unswappable allocator instances.

The atlas code needed some changes in the parsetree module, since it directly
constructs list nodes (without using |simple_list| or |sl_list|), and needs to
pass a valid (but currently stateless) deleter values in doing so.