Rename a local |struct raw| to |struct raw_tag| for cweavex convenience
Since cweavex is ignorant of scopes of definitions, it got confused by |raw|
being defined a class identifier in arithmetic.h but being used as an ordinary
identifier elsewhere. The simplest solution is to use a somewhat more elaborate
identifier for the local use in arithmetic.h, leaving |raw| a  untainted name.