Fix apparent error in |rat_value| definition: methods declared |&| should be |&&|
Thes are instances of methods |numerator| and |denominator|, overloaded on the
value type of |*this|. As for the internal class |big_rat| which |rat_value| is
a wrapper around, overloads should be for |const &| (then copy) and for |&&|
(then return rvalue ref), but the latter was specified |&| instead. Although no
actual harm was seemingly done, this surely must have been a clearical error.