Move |common_context| construction outside the full block constructor
It used to be that, contrary to the partial block contructor which takes a
|common_context| reference as argument, the full block constructor builds one
as a local variable, which disappears after the construction is completed. This
asymmetry is removed in favour of the partial constructor, so all callers to the
full block constructor now (also) need to build a |common_context| first, and
pass it (by refenerce) to the constructor. This prepares a future change in
which |common_block| will use the a |SubSystem| stored in the inner class.