Undo useless and confusing dualisation of |RootSystem| underlying a |SubSystem|
Although |RootSystem| does have methods that differ between a root system and
its dual, none of there are used for |SubSystem| objects. Therefore this change
in the manner the underlying |RootSystem| is generated does not require many
changes to existing code (just a few in Fokko), and avoids future confusion when
methods like |root| that redirect to the parent datum will be added.