Correct previous commit: forgot to change |common_block| full block constructor
With |rc| no longer referring to a contructor parameter, reference to |rc| in
the constructor's member initialisers now revert to one of those members (also
called |rc|), regardless of whether that member is already initialised or not.
(One would expect a good compiler catching the use a member's value before it is
initialised, but gcc does not do this.) The necessary change (a simplification)
was correctly applied in the partial block constructor, but the full block
constructor was forgotten. This has now been fixed.