Eliminate |mu_correction| as method, moving its code into |recursion_column|
Although concerning terms that are algebraically of a different nature, the
task the |mu_correction| did is only needed in the context, and at the end, of
the |recursion_column| computation. Therefore the call of this method causes
useless runtime overhead, in particular since several arguments have to be
passed to be able to continue wit already computed values. While there was some
documentation justification of having a separate method, the change can be made
while maintaining essentially the comments and presentation order. Therefore it
seems best to remove this legacy separation, for easier future maintenance.