Improve sorting functions: better merge, and exploit stable merge sort
It was possible to write |merge| exploiting the flexible while-loop syntax
(determining in each branch of a choice clause separately the loop continuation
condition). Also in |inv_standarisation| it was not necessary to spell out
sorting of pairs, whose second component gives the initial position, according
to their first component, as it was possible to instead call |sort| with a
comparison function that sorts position indexes according to the values at those
positions in the original array. Due to the fact that as implemented merge sort
gives a stable sort, the result is precisely what |inv_standarisation| returns.