Simplify sorting of simply linked lists
By introducing a new recursive (static) method |simple_list::sort_next|, the
need to form a vector of iterators for a list before sorting is eliminated, and
merge sorting can be done in a single forward pass over the list. The merging
proper is also concentrated in a static method |merge_aux| of |simple_list|,
thus avoiding source code duplication. (There is still some duplication between
|merge_aux| and the both |merge| methods that merge distinct lists together.)

Inside the merging routines, moving the link to the to-be-merged list into a
local variable that is moved back at the end might lead to some speeding up, by
avoiding repeated updates of the same memory location, although this is the kind
of optimisation that an optimising compiler could also find out and apply.