Make new |CharcterTable| constructor, and use it for type A tables
This constructor takes lists of conjugacy class representatives (with classical
class data and labels) and of irreducible characters (with classical irrep data
and labels) as arguments, and computes and stores renumbering information, so
that after-the-fact modification of character data (degrees and table) as was
previsously done is not necessary.

The construction of a conjugacy class representative for a partition was
rewritten, as a simple product of reflections with increasing indices works.