Reorganise W_characters.at for more flexibility and moderately more efficiency
The main constructor now takes a list of class representatives, and functions
for class size computation and membership testing as arguments, so that tailor
made versions can be used when available. The old constructor now calls this new
one with a generically generated list of representatives and generic functions
as arguments; since all existing calls still pass through the old constructor,
no performance gain is yet expected from this. However, the conjugacy tests have
been curried to be able to exploit the fact that the first argument is always
one of the class representatives, so that computations depending only on the
first argument can be done at table construction time. The conjugacy test
remains based on He's algorithm for finding minimal length representatives in
any elliptic conjugacy class; the implementation of this algorithm has been
somewhat streamlined (once a length decrease is sensed, work at the old level is
abandoned immediately and we proceed from a single shorter element; also within
a length level we proceed by distance to this initial element, in order to limit
equality testing to the current and previous distance).

The reduction to a common Levi (if it exists) where both elements to be compared
are relatively elliptic uses functions defined in conjugate.at rather than those
defined in W_characters.at itself, which appear strange and possibly not very
efficient, as the former are constructing root data for various Levi subgroups.