Improve efficiency of orbit generation in |Levi_normalizer_index| computation
Notably the lookup for previuously seen orbit elements (which are $k$-sets of
posroots) is sped up by encoding to integers and using binary search in the most
recent 2 levels generated rather than linear search on the whole orbit so far;
also element-by-element extension of the ful list is avoided. This is intended
to make conjugacy class size computation for (cases like) E8 less laborious.