Replace |Lie_type| field of |CharacterTable| by |WeylClassTable|
Since the main |CharacterTable| uses a |WeylClassTable|, it is a waste not to
retain this value in the produced |CharacterTable|. Apart from future
possibility for the |CharacterTable| to delegate work to the |WeylClassTable|,
this also makes it possible to access a |RootDatum| from the table. The main
reason that seems to have motivated the lowly |LieType| to be stored instead is
that combining them for different simple factors is easy, but it is better to
just solve this problem, with a product operation on |WeylClassTable| values;
this will ultimately be necessary anyway, so that is what has been added here.

The product operation assumes that coordinates of X^* remain the same, in other
words it is a product operation on Weyl groups of mutually orthogonal root
subsystems within the same space X^*.