Change class size argument to |W_class_table|: use class number rather than rep
Although a function computing the class size for an arbitrary Weyl group element
is natural to define, it is actually only used in the constructor for standard
class representatives, and callers often have extra information that can provide
the class size more easily than using a class representative. If the size is
computed from combinatorial data such as a cycle type, then that data is often
at hand since class representatives were computed using it. And for tables that
are composed from several simpler ones, as in character_tables_reductive.at,
this allows going straight for the sizes stored in the smaller tables, rather
than constructing composite Weyl group elements, decomposing, and looking up the
component Weyl group elements in each table to locate the correct stored size.