Fix defect |combine@(RootDatum,[CharacterTable])|: adapt to class reordering
The function used the list of class representatives from the |WeylClassTable| it
constructed for the composite group, but forgot to cater for the renumbering
that the |CharacterTable|s to be combined might invoke. Fields were added to
|CharacterTable| to make the necessary renumbering accessible externally, and
all |CharacterTable| constructing functions were adapted to it.

An alternative approach would have been to use the |class_reps| field from the
|CharacterTable| factors as source of Weyl group elements (using mixed radix
conversions) rather than |class_representatives()| from the combine class table.
That would be a simpler fix, but would mean that we need to convert those
elements, embedding them into the full Weyl group and multiplying these
together, which duplicates work that was done when constructing the combined
|WeylClassTable|. The current solution is therefore a bit more efficient, and
the added fields will probably prove necessary to have in other situations.