Avoid calls of |root_datum@(mat,mat)| that silently default |prefers_coroots|
We want the |prefers_coroots| to be inherited from the original root datum
wherever possible, which is almost everywhere. This ensures that if the full
root system gets selected, an identical root datum results. In a few cases there
is no root datum in the context, but even in those cases it is better to provide
the |prefers_coroots| explicitly even when it coincides with the default value,
so that the choice being made here is more explicit. In a few cases there is
more than one root datum in the context, in which case the first one rules.