Modify character_table_E8.at so that it takes into account a given root datum
It was the case that |character_table_E8@RootDatum| constructed a character
table for |simply_connected("E8")|, regardless of the precise properties of the
root datum passed. That is now changed, thought that root datum is still
privileged in that a class and charcter table for it are saved for re-use once
once a first instance has been called (whether providing that root datum as
explicit argument, or when calling |character_table_E8| without any argument.