Prepare obsolete data files for E8 to pass away in beauty
The files e8_magma.at and sym_powers_reflection_E8.at are not used (any more) to
build the character table for E8. Before removing them, they are cleaned up here
so that git will retain a cleaned-up version of them. Apart from formatting and
commenting, a reference to gap class size data in e8_magma.at was replaced,
which must have been an error since the ordering of classes is different,