Shorter version of GK_dimension
Existing code for GK_dimension@Param does a lot of calculation in the
character table for G to get a nilpotent for G; but the desired GK
dimension was known from the calculation down in the integral root
datum. This version "GKfast@Param" cuts to the chase. Should be
noticeably faster for large G and nonintegral infinitesimal
character. Should be incorporated into
associated_variety_annihilator.at