Simplify character_tables.at, removing stuff that is no longer necessary
A |CharacterTable| and its contained |WeylClassTable| are built for (want with
Weyl group elements of) the same root datum, so it is not useful to distinguish
a |parent_root_datum|. Also functions like |update_degrees| are no longer used
since the (fake) degrees are computed at construction. Thus character_tables.at
was cleaned up, and other scripts adapted to use the simplified interface.