Rationalize structure of |WeylClassTable| and consequently of |CharacterTable|.
The main improvement is implementing a reordering function from a
|WeylClassTable| to a distinct |WeylClassTable| with imposed order of the
classes. This moves functionality out of the |CharacterTable| constructor (since
it just reorders the class table before using it), which makes the renumbering
member of that type obsolete. Then also other members can be handled directly by
the contained |WeylClassTable|, and can be removed from |WeylClassTable| which
ends up being greakly simplified.