Remove and shuflle some code, reducing triplicate W-action code
This shortens character_tables.at even more, as several functions with
|CaharacterTable| argument can be handled directly by their |WeylClassTable|
counterpart. The construction of matrices for a W-action (for instance defined
by a cell) was present 3 times; it was maintained (with |WeylClassTable|
argument) in W_reps.at, removed both from character_tables.at and from
character_table_reps.at. A small change was made to computing dimension of a
character (which is just evaluation at the identity class); the version
|dimension@CharacterTable| is no a function ([int]->int) that computes it for a
general character, while the computation for an irreducible from the table,
identied by its index, is now |dimension@(CharacterTable,int)| (as it could not
reamin in its curried form: only one |dimension@CharacterTable| can exist).