Fix definitions of |simply_connected@RootDatum| and |adjoint@RootDatum|
The former just used the |Lie_type| of its argument, which ignores both the
ordering in the Cartan matrix (numbering of the diagram) and the status with
respect to |prefers_coroots| which governs the order in which non-simple roots
will be present in the root datum. The latter had two definitions, one with the
same problems as |simply_connected@RootDatum|, and another which by using the
Cartan matrix solved the first problem, but overrided the |prefers_coroots|
status of the given root datum by an explict |false|: this cuased a mismatch in
the ordering of the non-simple roots in the presence of different root lengths.