Group all useful forms of |sub_datum| in basic.at, adapt W_characters.at
There are basically two main ways to form a |sub_datum| of a given one: specify
exactly the indices in order of the new simple roots, or specify a condition for
the (positive) roots to retain. The former is the most general in that it can
use negative roots for the new simple roots, but the user has the task of
prepare a proper list of root indices; the latter is often sufficient, and can
do the work of reducing a positive system to a simple one, but it will always
retain positivity and the relative order of the selected simple roots. The
existing |sub_datum@(RootDatum,[int])| was only intended for lists of indices
of already simple roots, and has been generalised to handle more general lists.
The latter was already present as |Levi_datum| in W_characters.at, and that
definition has now become an alias for one form basic.at