Give character table lazily updated generic (previously fake) degree component
Instead of having as component of |CharacterTable| a simple list of generic
degrees that may or may not have been computed, with |update_fake_degrees|
replacing that component by a more definitive list, we now have an internal
private table of generic degrees, accessed by three functions: the predicate
|has_generic_degree| tells whether a specific degree has been set, the function
|generic_degree| returns that degree provided it was set (and gives an error if
not), while |set_generic_degree| defines a degree if it was not yet previously,
optionally testing the provided value in case it was already present (in order
to not do any work in this case if no test is opted for, the degree to set to is
passed as a function without arguments, so that is need only be called if
actually needed). This is a half-way solution (since |CharacterTable| is still
dependent on external actions to actually provide the generic degrees) that
allows adapting the current work-flow with minimal changes; ultimately we would
like to make this lazy with on-demand computation of generic degrees.